Use error handling from apollo-server; throwError should only be defined on server
