+++ packages/vulcan-lib/lib/server/main.js
@@ -8,6 +8,7 @@ export * from './connectors.js';
 export * from './query.js';
 export * from '../modules/index.js';
 export * from './mutators.js';
+export * from './errors.js';
 // TODO: what to do with this?
 export * from './meteor_patch.js';
 //export * from './render_context.js';
